+++ sdk/search/Azure.Search.Documents/tests/DocumentOperations/VectorSearchTests.cs
@@ -153,31 +153,6 @@ namespace Azure.Search.Documents.Tests
         }
 
         [Test]
-        [PlaybackOnly("The availability of Semantic Search is limited to specific regions, as indicated in the list provided here: https://azure.microsoft.com/explore/global-infrastructure/products-by-region/?products=search. Due to this limitation, the deployment of resources for weekly test pipeline for setting the \"semanticSearch\": \"free\" fails in the UsGov and China cloud regions.")]
-        public async Task SemanticMaxWaitOutOfRangeThrows()
-        {
-            await using SearchResources resources = await SearchResources.CreateWithHotelsIndexAsync(this);
-
-            RequestFailedException ex = await CatchAsync<RequestFailedException>(
-                async () => await resources.GetSearchClient().SearchAsync<Hotel>(
-                    "Is there any luxury hotel in New York?",
-                    new SearchOptions
-                    {
-                        SemanticSearch = new()
-                        {
-                            SemanticConfigurationName = "my-semantic-config",
-                            QueryCaption = new(QueryCaptionType.Extractive),
-                            QueryAnswer = new(QueryAnswerType.Extractive),
-                            MaxWait = TimeSpan.FromMilliseconds(700),
-                        },
-                        QueryType = SearchQueryType.Semantic,
-                    }));
-
-            Assert.AreEqual(400, ex.Status);
-            Assert.AreEqual("InvalidRequestParameter", ex.ErrorCode);
-        }
-
-        [Test]
         public async Task UpdateExistingIndexToAddVectorFields()
         {
             await using SearchResources resources = SearchResources.CreateWithNoIndexes(this);